iot-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.15.0
sqlalchemy>=2.0.0
alembic>=1.16.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=1.1.3 